added error handling

+++ pkg/vmo/deployment.go
@@ -42,7 +42,11 @@ func updateOpenSearchDashboardsDeployment(osd *appsv1.Deployment, controller *Co
 		err = updateDeployment(controller, vmo, existingDeployment, osd)
 	}
 	if err != nil {
-		metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
+		metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+		if metricErr != nil {
+			return metricErr
+		}
+		metric.Inc()
 		controller.log.Errorf("Failed to update deployment %s/%s: %v", osd.Namespace, osd.Name, err)
 		return err
 	}
@@ -96,7 +100,11 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 			}
 		}
 		if err != nil {
-			metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
+			metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+			if metricErr != nil {
+				return false, metricErr
+			}
+			metric.Inc()
 			controller.log.Errorf("Failed to update deployment %s/%s: %v", curDeployment.Namespace, curDeployment.Name, err)
 			return false, err
 		}
@@ -144,18 +152,30 @@ func CreateDeployments(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMon
 }
 
 func deleteDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, deployment *appsv1.Deployment) error {
-	metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentDeleteCounter).Inc()
+	metric, metricErr := metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentDeleteCounter)
+	if metricErr != nil {
+		return metricErr
+	}
+	metric.Inc()
 	controller.log.Debugf("Deleting deployment %s", deployment.Name)
 	err := controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Delete(context.TODO(), deployment.Name, metav1.DeleteOptions{})
 	if err != nil {
-		metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentDeleteError).Inc()
+		metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentDeleteError)
+		if metricErr != nil {
+			return metricErr
+		}
+		metric.Inc()
 		controller.log.Errorf("Failed to delete deployment %s: %v", deployment.Name, err)
 	}
 	return nil
 }
 
 func updateDeployment(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitoringInstance, existingDeployment, curDeployment *appsv1.Deployment) error {
-	metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
+	metric, metricErr := metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+	if metricErr != nil {
+		return metricErr
+	}
+	metric.Inc()
 	var err error
 	curDeployment.Spec.Selector = existingDeployment.Spec.Selector
 	specDiffs := diff.Diff(existingDeployment, curDeployment)
@@ -182,7 +202,11 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 		if !isUpdateAllowed(controller, vmo, current) {
 			continue
 		}
-		metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
+		metric, metricErr := metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+		if metricErr != nil {
+			return false, metricErr
+		}
+		metric.Inc()
 		// Selector may not change, so we copy over from existing
 		current.Spec.Selector = existing.Spec.Selector
 		// Deployment spec differences, so call Update() and return
@@ -192,7 +216,11 @@ func rollingUpdate(controller *Controller, vmo *vmcontrollerv1.VerrazzanoMonitor
 			controller.log.Oncef("Updating deployment %s in namespace %s", current.Name, current.Namespace)
 			_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), current, metav1.UpdateOptions{})
 			if err != nil {
-				metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
+				metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+				if metricErr != nil {
+					return false, metricErr
+				}
+				metric.Inc()
 				return false, err
 			}
 			//okay to return dirty=false after updating the *last* deployment
@@ -222,10 +250,18 @@ func updateAllDeployments(controller *Controller, vmo *vmcontrollerv1.Verrazzano
 		if err != nil {
 			return false, err
 		}
-		metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter).Inc()
+		metric, metricErr := metricsexporter.GetSimpleCounterMetrics(metricsexporter.NamesDeploymentUpdateCounter)
+		if metricErr != nil {
+			return false, metricErr
+		}
+		metric.Inc()
 		_, err = controller.kubeclientset.AppsV1().Deployments(vmo.Namespace).Update(context.TODO(), curDeployment, metav1.UpdateOptions{})
 		if err != nil {
-			metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError).Inc()
+			metric, metricErr := metricsexporter.GetErrorMetrics(metricsexporter.NamesDeploymentUpdateError)
+			if metricErr != nil {
+				return false, metricErr
+			}
+			metric.Inc()
 			return false, err
 		}
 	}
